feat(utils): add StatSchema
