+++ packages/utils/schema/lowerthird.schema.ts
@@ -1,24 +1,5 @@
-import { string, z } from "zod"
-
-export const adjTextSchema = z.object({
-  text: z.string(),
-  size: z.number().default(0),
-})
-
-export const adjSizeSchema = z.object({
-  h: z.number().default(0),
-  w: z.number().default(0),
-  x: z.number().default(0),
-  y: z.number().default(0),
-  scale: z.number().default(0),
-})
-
-export const adjImageSchema = z.object({
-  URL: z.string(),
-  BASE64: z.string().optional(),
-  adj: adjSizeSchema,
-})
-
+import { z } from "zod"
+import { adjImageSchema, adjSizeSchema, adjTextSchema } from "./adj.schema"
 // --------
 
 export const tickerSchema = z.object({
